Conflicts:
	.gitignore
	HotelReservationSystemClient/src/java/hotelreservationsystemclient/MainApp.java
	HotelReservationSystemLibrary/src/entity/Guest.java
	HotelReservationSystemLibrary/src/entity/Transaction.java
	HotelReservationSystemLibrary/src/util/exception/UnknownPersistenceException.java

+++ HotelReservationSystem-ejb/src/java/ejb/session/stateless/ReservationSessionBeanLocal.java
@@ -18,6 +18,8 @@ import util.exception.UnknownPersistenceException;
  */
 @Local
 public interface ReservationSessionBeanLocal {
-    public BigDecimal calculatePrice(LocalDateTime checkInDate, LocalDateTime checkOutDate, RoomType roomType, String reservationType);
-     public Long createNewOnlineReservation(Reservation reservation, Long roomTypeId, Long guestId) throws UnknownPersistenceException;
+
+    public Double calculatePrice(LocalDateTime checkInDate, LocalDateTime checkOutDate, RoomType roomType, String reservationType);
+
+    public Long createNewOnlineReservation(Reservation reservation, Long roomTypeId, Long guestId) throws UnknownPersistenceException;
 }
